+++ test/e2e/subscriber_e2e_test.dart
@@ -44,6 +44,9 @@ class _FakePushTokenService implements PushTokenService {
 
   @override
   Future<void> openSettings() async {}
+
+  @override
+  Future<ApnsEnvironment?> detectApnsEnvironment() async => null;
 }
 
 void main() {